+++ src/main/java/org/thingai/app/controller/ScoreController.java
@@ -3,14 +3,15 @@ package org.thingai.app.controller;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.thingai.app.controller.utils.ResponseEntityUtil;
 import org.thingai.app.scoringservice.ScoringService;
 import org.thingai.app.scoringservice.callback.RequestCallback;
 import org.thingai.app.scoringservice.entity.score.Score;
-import org.thingai.app.controller.utils.ResponseEntityUtil;
 import org.thingai.app.scoringservice.entity.score.ScoreDefine;
 
 import java.util.HashMap;
 import java.util.concurrent.CompletableFuture;
+
 import static org.thingai.app.controller.utils.ResponseEntityUtil.createErrorResponse;
 
 @RestController
@@ -85,7 +86,7 @@ public class ScoreController {
     @GetMapping("/ui-definitions")
     public ResponseEntity<Object> getScoreUIDefinitions() {
         CompletableFuture<ResponseEntity<Object>> future = new CompletableFuture<>();
-        ScoringService.scoreHandler().getScoreUi(new RequestCallback<HashMap<String, ScoreDefine>>() {
+        ScoringService.scoreHandler().getScoreDefinitions(new RequestCallback<HashMap<String, ScoreDefine>>() {
             @Override
             public void onSuccess(HashMap<String, ScoreDefine> responseObject, String message) {
                 future.complete(ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(responseObject));
